improve example server

+++ example_server.ts
@@ -5,10 +5,20 @@ import fs from "node:fs";
 import path from "node:path";
 
 const dest_dir = "target";
-const example_module_name = process.argv.pop();
+const example_module_name = process.argv.pop() ?? "";
 if (!example_module_name || example_module_name.includes("example_server.js")) {
   console.log("pass an example name to run. Example: npm run example shape_2d");
   process.exit(1);
+}
+
+// checks If it is print cmd
+if (
+  example_module_name === "--help"
+  || example_module_name.includes("example_server.ts")
+  || example_module_name === ""
+) {
+  await print_help();
+  process.exit(0);
 }
 
 async function run_command(cmd: string, options = {}): Promise<void> {
@@ -23,7 +33,35 @@ async function run_command(cmd: string, options = {}): Promise<void> {
   });
 }
 
+async function get_available_example_names(): Promise<string[]> {
+  const dir_path = path.resolve("examples");
+  const files = fs.readdirSync(dir_path);
+  const ignoreFiles = ["res", "main.ts"];
+
+  return files
+    .filter(name => !ignoreFiles.includes(name))
+    .map(name => name.replace(".ts", ""));
+}
+
+async function print_help(): Promise<void> {
+  const available_examples = await get_available_example_names();
+  console.log("Usage:");
+  console.log("npm run example <example_name>");
+  console.log("\nAvailable example names:")
+  console.log("-", available_examples.join("\n- "));
+}
+
 async function main(): Promise<void> {
+  const available_examples = await get_available_example_names();
+
+  if (!available_examples.includes(example_module_name)) {
+    console.log(`Example: '${example_module_name}' does not exist!`);
+    console.log(`Below are the available example names:\n-`, available_examples.join("\n- "));
+    process.exit(0);
+  }
+
+  console.log(`running example: ${example_module_name}`);
+
   try {
     if (fs.existsSync(dest_dir)) {
       fs.rmSync(dest_dir, { recursive: true, force: true });
